+++ vrep-control-loop/src/de/joachim/haensel/phd/scenario/vehicle/navigation/ITrajectorizer.java
@@ -7,5 +7,5 @@ import de.joachim.haensel.sumo2vrep.Line2D;
 public interface ITrajectorizer
 {
     public List<Trajectory> createTrajectory(List<Line2D> downscaled);
-    public float[][] getPoints();
+    public double[][] getPoints();
 }
